Fixed problem with blacklisting and comments.
Fixed problem with blacklisting, when the last added blacklist comment appeared in blacklistmodal when blacklisting package and with comments, when the comment was not saved and you switched the package version(comment remained the same)
