+++ ota-plus-web/app/reactapp/src/components/device/packages/index.js
@@ -3,11 +3,13 @@ import PackagesListItem from './ListItem';
 import PackagesListItemVersion from './ListItemVersion';
 import PackagesListItemOnDevice from './ListItemOnDevice';
 import PackagesDetails from './Details';
+import PackagesComment from './Comment';
 
 export {
     PackagesList,
     PackagesListItem,
     PackagesListItemVersion,
     PackagesListItemOnDevice,
-    PackagesDetails
+    PackagesDetails,
+    PackagesComment
 }
